% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_length.R
\name{analyse_length}
\alias{analyse_length}
\title{Analyse length-frequency data from annual VEFMAP surveys}
\usage{
analyse_length(x, species, years, col, ...)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to analyse length-frequency data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
