% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_recruits.R
\name{analyse_recruits}
\alias{analyse_recruits}
\title{Analyse recruitment data from annual VEFMAP surveys}
\usage{
analyse_recruits(x, species, years, col, ...)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to analyse recruitment data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
