% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_trends.R
\name{analyse_trends}
\alias{analyse_trends}
\title{Analyse catch data from annual VEFMAP surveys}
\usage{
analyse_trends(x, species, years, col, ...)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to analyse catch data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
