% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean survey data associated with annual VEFMAP surveys}
\usage{
clean_data(x, systems = NULL, species = NULL, years = NULL, gear = NULL)
}
\arguments{
\item{x}{djf}

\item{systems}{df}

\item{species}{df}

\item{years}{sf}

\item{gear}{d}
}
\description{
Series of functions to clean the data from VEFMAP surveys
and standardise all column names and values
}
\details{

}
\examples{



}
