% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_distribution.R
\name{examine_distribution}
\alias{examine_distribution}
\title{Examine distributions of species at sites in annual VEFMAP surveys}
\usage{
examine_distribution(
  x,
  survey,
  system,
  species = NULL,
  reach = NULL,
  start = NULL,
  end = NULL,
  ...
)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to examine distributions of species at sites
from VEFMAP surveys and define a class based on these outputs
}
\details{

}
\examples{



}
