% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_hydro.R
\name{examine_hydro}
\alias{examine_hydro}
\title{Examine hydrological data associated with annual VEFMAP surveys}
\usage{
examine_hydro(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{djf}

\item{start}{df}

\item{end}{df}

\item{y}{ignored, for consistency with S3 plot method}
}
\description{
Series of functions to examine hydrological data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
