% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_length.R
\name{examine_length}
\alias{examine_length}
\title{Examine length-frequency data from annual VEFMAP surveys}
\usage{
examine_length(x, species, system, reach = NULL, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to examine length-frequency data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
