% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_species.R
\name{examine_species}
\alias{examine_species}
\title{Return species table from VEFMAP data}
\usage{
examine_species(x)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to analyse length-frequency data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
