% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_surveys.R
\name{examine_surveys}
\alias{examine_surveys}
\title{Return survey table from VEFMAP data}
\usage{
examine_surveys(x)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to analyse length-frequency data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
