% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_trend.R
\name{examine_trend}
\alias{examine_trend}
\title{Examine catch data from annual VEFMAP surveys}
\usage{
examine_trend(x, species, system, reach = NULL, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{}
}
\description{
Series of functions to examine catch data from VEFMAP surveys
and define a class based on these outputs
}
\details{

}
\examples{



}
